+++ src/main/java/woowacourse/exception/GlobalControllerAdvice.java
@@ -1,33 +1,78 @@
 package woowacourse.exception;
 
+import java.util.List;
+import javax.validation.ConstraintViolationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.HttpMessageNotReadableException;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import woowacourse.auth.exception.UnAuthorizedException;
 import woowacourse.exception.dto.ErrorResponse;
 import woowacourse.member.exception.BadRequestException;
+import woowacourse.shoppingcart.exception.InvalidCartItemException;
+import woowacourse.shoppingcart.exception.InvalidCustomerException;
+import woowacourse.shoppingcart.exception.InvalidOrderException;
+import woowacourse.shoppingcart.exception.InvalidProductException;
+import woowacourse.shoppingcart.exception.NotInCustomerCartItemException;
 
 @RestControllerAdvice
 public class GlobalControllerAdvice {
 
-    private final Logger logger;
-
-    public GlobalControllerAdvice() {
-        logger = LoggerFactory.getLogger(this.getClass());
-    }
+    private static final Logger LOGGER = LoggerFactory.getLogger(GlobalControllerAdvice.class);
 
     @ExceptionHandler(BadRequestException.class)
     public ResponseEntity<ErrorResponse> badRequestExceptionHandler(BadRequestException e) {
-        logger.error(e.getMessage(), e);
+        LOGGER.error(e.getMessage(), e);
         return ResponseEntity.badRequest().body(new ErrorResponse(e.getMessage()));
     }
 
     @ExceptionHandler(UnAuthorizedException.class)
     public ResponseEntity<ErrorResponse> unAuthorizedExceptionHandler(UnAuthorizedException e) {
-        logger.error(e.getMessage(), e);
+        LOGGER.error(e.getMessage(), e);
         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse(e.getMessage()));
+    }
+
+    @ExceptionHandler(RuntimeException.class)
+    public ResponseEntity handleUnhandledException() {
+        return ResponseEntity.badRequest().body("Unhandled Exception");
+    }
+
+    @ExceptionHandler(EmptyResultDataAccessException.class)
+    public ResponseEntity handle() {
+        return ResponseEntity.badRequest().body("존재하지 않는 데이터 요청입니다.");
+    }
+
+    @ExceptionHandler({MethodArgumentNotValidException.class})
+    public ResponseEntity handleInvalidRequest(final BindingResult bindingResult) {
+        final List<FieldError> fieldErrors = bindingResult.getFieldErrors();
+        final FieldError mainError = fieldErrors.get(0);
+
+        return ResponseEntity.badRequest().body(mainError.getDefaultMessage());
+    }
+
+    @ExceptionHandler({
+            HttpMessageNotReadableException.class,
+            ConstraintViolationException.class,
+    })
+    public ResponseEntity handleInvalidRequest(final RuntimeException e) {
+        return ResponseEntity.badRequest().body(e.getMessage());
+    }
+
+    @ExceptionHandler({
+            InvalidCustomerException.class,
+            InvalidCartItemException.class,
+            InvalidProductException.class,
+            InvalidOrderException.class,
+            NotInCustomerCartItemException.class,
+    })
+    public ResponseEntity handleInvalidAccess(final RuntimeException e) {
+        return ResponseEntity.badRequest().body(e.getMessage());
     }
 }
